Use gm_unsplit helper header
