+++ src/melee/gm/gm_1832.h
@@ -1,0 +1,75 @@
+#ifndef MELEE_GM_1832_H
+#define MELEE_GM_1832_H
+
+#include <placeholder.h>
+#include <sysdolphin/baselib/forward.h>
+
+/* 1849E0 */ void fn_801849E0(HSD_GObj*);
+/* 184A04 */ void fn_80184A04(HSD_GObj*);
+/* 184A28 */ void fn_80184A28(HSD_GObj*);
+/* 184A4C */ void fn_80184A4C(HSD_GObj*);
+/* 184A70 */ void fn_80184A70(HSD_GObj*);
+/* 184A94 */ void fn_80184A94(HSD_GObj*);
+/* 184AB8 */ void fn_80184AB8(HSD_GObj*);
+/* 18504C */ UNK_RET fn_8018504C(UNK_PARAMS);
+/* 1851C0 */ UNK_RET fn_801851C0(UNK_PARAMS);
+/* 1852FC */ void fn_801852FC(HSD_GObj*);
+/* 185408 */ void fn_80185408(int, float, float, float, float);
+/* 1855BC */ double fn_801855BC(double);
+/* 18564C */ void fn_8018564C(HSD_GObj*);
+/* 18569C */ void fn_8018569C(HSD_GObj*);
+/* 18575C */ void fn_8018575C(HSD_GObj*);
+/* 1857C4 */ void fn_801857C4(HSD_GObj*);
+/* 1859C8 */ UNK_RET fn_801859C8(UNK_PARAMS);
+/* 185A0C */ UNK_RET fn_80185A0C(UNK_PARAMS);
+/* 185D64 */ UNK_RET fn_80185D64(UNK_PARAMS);
+/* 185E34 */ UNK_RET fn_80185E34(UNK_PARAMS);
+/* 185F5C */ UNK_RET fn_80185F5C(UNK_PARAMS);
+/* 186080 */ UNK_RET fn_80186080(UNK_PARAMS);
+/* 1861B8 */ UNK_RET fn_801861B8(UNK_PARAMS);
+/* 186400 */ UNK_RET fn_80186400(UNK_PARAMS);
+/* 186634 */ UNK_RET fn_80186634(UNK_PARAMS);
+/* 186DFC */ UNK_RET gm_80186DFC_OnFrame(UNK_PARAMS);
+/* 186E30 */ void gm_80186E30_OnEnter(UNK_T);
+/* 186EFC */ UNK_RET fn_80186EFC(UNK_PARAMS);
+/* 186F6C */ UNK_RET fn_80186F6C(UNK_PARAMS);
+/* 1873F0 */ UNK_RET fn_801873F0(UNK_PARAMS);
+/* 187494 */ void fn_80187494(HSD_GObj*, int);
+/* 1874FC */ UNK_RET fn_801874FC(UNK_PARAMS);
+/* 187714 */ UNK_RET fn_80187714(UNK_PARAMS);
+/* 18776C */ UNK_RET gm_8018776C_OnFrame(UNK_PARAMS);
+/* 1877A8 */ void gm_801877A8_OnEnter(UNK_T);
+/* 187910 */ UNK_RET fn_80187910(UNK_PARAMS);
+/* 187AB4 */ UNK_RET fn_80187AB4(UNK_PARAMS);
+/* 187C9C */ UNK_RET fn_80187C9C(UNK_PARAMS);
+/* 187CF4 */ UNK_RET fn_80187CF4(UNK_PARAMS);
+/* 187F48 */ void gm_80187F48_OnEnter(UNK_T);
+/* 188364 */ void gm_80188364_OnLeave(UNK_T);
+/* 18838C */ UNK_RET gm_8018838C_OnFrame(UNK_PARAMS);
+/* 1883C0 */ UNK_RET gm_801883C0(UNK_PARAMS);
+/* 18841C */ bool gm_8018841C(void);
+/* 188454 */ int gm_80188454(int idx);
+/* 18846C */ UNK_RET fn_8018846C(UNK_PARAMS);
+/* 1884F8 */ UNK_RET fn_801884F8(UNK_PARAMS);
+/* 188550 */ UNK_RET fn_80188550(UNK_PARAMS);
+/* 188644 */ UNK_RET fn_80188644(UNK_PARAMS);
+/* 188738 */ UNK_RET fn_80188738(UNK_PARAMS);
+/* 188910 */ UNK_RET fn_80188910(UNK_PARAMS);
+/* 188B3C */ UNK_RET fn_80188B3C(UNK_PARAMS);
+/* 188D3C */ UNK_RET fn_80188D3C(UNK_PARAMS);
+/* 188EE8 */ UNK_RET fn_80188EE8(UNK_PARAMS);
+/* 1891F4 */ UNK_RET fn_801891F4(UNK_PARAMS);
+/* 189B88 */ UNK_RET fn_80189B88(UNK_PARAMS);
+/* 189CDC */ UNK_RET gm_80189CDC(UNK_PARAMS);
+/* 18A000 */ UNK_RET fn_8018A000(UNK_PARAMS);
+/* 18A160 */ UNK_RET gm_8018A160(UNK_PARAMS);
+/* 18A188 */ UNK_RET gm_8018A188(UNK_PARAMS);
+/* 18A1D8 */ UNK_RET gm_8018A1D8(UNK_PARAMS);
+/* 18A228 */ UNK_RET gm_8018A228(UNK_PARAMS);
+/* 18A25C */ UNK_RET gm_8018A25C(UNK_PARAMS);
+/* 18A290 */ UNK_RET gm_8018A290(UNK_PARAMS);
+/* 18A2C4 */ UNK_RET gm_8018A2C4(UNK_PARAMS);
+/* 18A314 */ UNK_RET gm_8018A314(UNK_PARAMS);
+/* 18A364 */ void fn_8018A364(int);
+
+#endif
