+++ src/melee/gm/gm_18A5.h
@@ -1,0 +1,139 @@
+#ifndef MELEE_GM_18A5_H
+#define MELEE_GM_18A5_H
+
+#include <placeholder.h>
+#include <melee/gm/forward.h>
+#include <sysdolphin/baselib/forward.h>
+
+/* 18A514 */ UNK_RET fn_8018A514(UNK_PARAMS);
+/* 18A970 */ UNK_RET fn_8018A970(UNK_PARAMS);
+/* 18AA74 */ UNK_RET fn_8018AA74(UNK_PARAMS);
+/* 18B090 */ UNK_RET fn_8018B090(UNK_PARAMS);
+/* 18C8D4 */ UNK_RET fn_8018C8D4(UNK_PARAMS);
+/* 18D50C */ UNK_RET fn_8018D50C(UNK_PARAMS);
+/* 18DC18 */ UNK_RET fn_8018DC18(UNK_PARAMS);
+/* 18DF68 */ UNK_RET fn_8018DF68(UNK_PARAMS);
+/* 18E46C */ UNK_RET fn_8018E46C(UNK_PARAMS);
+/* 18E618 */ UNK_RET fn_8018E618(UNK_PARAMS);
+/* 18E85C */ UNK_RET fn_8018E85C(UNK_PARAMS);
+/* 18EC48 */ UNK_RET fn_8018EC48(UNK_PARAMS);
+/* 18EC7C */ UNK_RET fn_8018EC7C(UNK_PARAMS);
+/* 18ECA8 */ UNK_RET fn_8018ECA8(UNK_PARAMS);
+/* 18F00C */ UNK_RET fn_8018F00C(UNK_PARAMS);
+/* 18F1B0 */ UNK_RET gm_8018F1B0(UNK_PARAMS);
+/* 17F294 */ int fn_8017F294(void);
+/* 18F310 */ UNK_RET fn_8018F310(UNK_PARAMS);
+/* 18F3BC */ u8 fn_8018F3BC(s32 arg0);
+/* 18F3D0 */ UNK_RET fn_8018F3D0(UNK_PARAMS);
+/* 18F410 */ UNK_RET fn_8018F410(UNK_PARAMS);
+/* 18F4A0 */ UNK_RET fn_8018F4A0(UNK_PARAMS);
+/* 18F508 */ UNK_RET fn_8018F508(UNK_PARAMS);
+/* 18F5F0 */ char* fn_8018F5F0(UNK_PARAMS);
+/* 18F62C */ s32 fn_8018F62C(HSD_GObj*);
+/* 18F634 */ TmData* gm_8018F634(void);
+/* 18F640 */ UNK_RET fn_8018F640(UNK_PARAMS);
+/* 18F674 */ UNK_RET fn_8018F674(UNK_PARAMS);
+/* 18F6A8 */ UNK_RET fn_8018F6A8(UNK_PARAMS);
+/* 18F6DC */ UNK_RET fn_8018F6DC(UNK_PARAMS);
+/* 18F6FC */ UNK_RET fn_8018F6FC(UNK_PARAMS);
+/* 18F71C */ UNK_RET fn_8018F71C(UNK_PARAMS);
+/* 18F74C */ UNK_RET fn_8018F74C(UNK_PARAMS);
+/* 18F808 */ UNK_RET fn_8018F808(UNK_PARAMS);
+/* 18F888 */ UNK_RET fn_8018F888(UNK_PARAMS);
+/* 18FA24 */ UNK_RET fn_8018FA24(UNK_PARAMS);
+/* 18FBD8 */ void fn_8018FBD8(UNK_T arg0, s32 arg1);
+/* 18FBE0 */ UNK_RET fn_8018FBE0(UNK_PARAMS);
+/* 18FDC4 */ void fn_8018FDC4(HSD_JObj*, float, float, float);
+/* 18FF9C */ UNK_RET fn_8018FF9C(UNK_PARAMS);
+/* 190174 */ UNK_RET fn_80190174(UNK_PARAMS);
+/* 1901F8 */ UNK_RET fn_801901F8(UNK_PARAMS);
+/* 19027C */ UNK_RET fn_8019027C(UNK_PARAMS);
+/* 1902F0 */ UNK_RET fn_801902F0(UNK_PARAMS);
+/* 19035C */ UNK_RET fn_8019035C(UNK_PARAMS);
+/* 19044C */ void fn_8019044C(HSD_JObj*, float);
+/* 190480 */ UNK_RET fn_80190480(UNK_PARAMS);
+/* 1904D0 */ UNK_RET fn_801904D0(UNK_PARAMS);
+/* 190520 */ UNK_RET fn_80190520(UNK_PARAMS);
+/* 1905F0 */ UNK_RET gm_801905F0(UNK_PARAMS);
+/* 190ABC */ UNK_RET fn_80190ABC(UNK_PARAMS);
+/* 190EA4 */ UNK_RET gm_80190EA4(UNK_PARAMS);
+/* 190FE4 */ UNK_RET gm_80190FE4(UNK_PARAMS);
+/* 1910E0 */ UNK_RET fn_801910E0(UNK_PARAMS);
+/* 191154 */ UNK_RET fn_80191154(UNK_PARAMS);
+/* 191240 */ UNK_RET fn_80191240(UNK_PARAMS);
+/* 1913BC */ UNK_RET fn_801913BC(UNK_PARAMS);
+/* 191678 */ UNK_RET fn_80191678(UNK_PARAMS);
+/* 19175C */ UNK_RET fn_8019175C(UNK_PARAMS);
+/* 1918F0 */ UNK_RET fn_801918F0(UNK_PARAMS);
+/* 191A54 */ UNK_RET fn_80191A54(UNK_PARAMS);
+/* 191B5C */ UNK_RET fn_80191B5C(UNK_PARAMS);
+/* 191CA4 */ UNK_RET fn_80191CA4(UNK_PARAMS);
+/* 191D38 */ UNK_RET fn_80191D38(UNK_PARAMS);
+/* 191E9C */ UNK_RET fn_80191E9C(UNK_PARAMS);
+/* 191FD4 */ UNK_RET fn_80191FD4(UNK_PARAMS);
+/* 19237C */ UNK_RET fn_8019237C(UNK_PARAMS);
+/* 19249C */ UNK_RET fn_8019249C(UNK_PARAMS);
+/* 192690 */ UNK_RET fn_80192690(UNK_PARAMS);
+/* 192758 */ UNK_RET fn_80192758(UNK_PARAMS);
+/* 192938 */ UNK_RET fn_80192938(UNK_PARAMS);
+/* 192BB0 */ UNK_RET fn_80192BB0(UNK_PARAMS);
+/* 192E6C */ UNK_RET fn_80192E6C(UNK_PARAMS);
+/* 193230 */ UNK_RET fn_80193230(UNK_PARAMS);
+/* 193308 */ UNK_RET fn_80193308(UNK_PARAMS);
+/* 1935B8 */ UNK_RET fn_801935B8(UNK_PARAMS);
+/* 1937C4 */ UNK_RET fn_801937C4(UNK_PARAMS);
+/* 193B58 */ UNK_RET fn_80193B58(UNK_PARAMS);
+/* 193FCC */ UNK_RET fn_80193FCC(UNK_PARAMS);
+/* 194658 */ UNK_RET fn_80194658(UNK_PARAMS);
+/* 1949B4 */ UNK_RET fn_801949B4(UNK_PARAMS);
+/* 194BC4 */ UNK_RET fn_80194BC4(UNK_PARAMS);
+/* 194D84 */ UNK_RET fn_80194D84(UNK_PARAMS);
+/* 194F30 */ UNK_RET fn_80194F30(UNK_PARAMS);
+/* 1953C8 */ UNK_RET fn_801953C8(UNK_PARAMS);
+/* 195AF0 */ UNK_RET fn_80195AF0(UNK_PARAMS);
+/* 195CCC */ UNK_RET fn_80195CCC(UNK_PARAMS);
+/* 19610C */ UNK_RET fn_8019610C(UNK_PARAMS);
+/* 19628C */ void gm_8019628C_OnFrame(void);
+/* 1963B4 */ void gm_801963B4_OnEnter(UNK_T);
+/* 1964A4 */ void gm_801964A4_OnLeave(UNK_T);
+/* 196510 */ UNK_RET fn_80196510(UNK_PARAMS);
+/* 19655C */ s32 fn_8019655C(void);
+/* 196564 */ UNK_RET fn_80196564(UNK_PARAMS);
+/* 196594 */ UNK_RET fn_80196594(UNK_PARAMS);
+/* 1965C4 */ UNK_RET fn_801965C4(UNK_PARAMS);
+/* 196684 */ UNK_RET fn_80196684(UNK_PARAMS);
+/* 1967E0 */ UNK_RET fn_801967E0(UNK_PARAMS);
+/* 196CF8 */ UNK_RET fn_80196CF8(UNK_PARAMS);
+/* 196DBC */ UNK_RET fn_80196DBC(UNK_PARAMS);
+/* 196E30 */ UNK_RET fn_80196E30(UNK_PARAMS);
+/* 196EEC */ UNK_RET fn_80196EEC(UNK_PARAMS);
+/* 196FFC */ UNK_RET fn_80196FFC(UNK_PARAMS);
+/* 1973F8 */ UNK_RET fn_801973F8(UNK_PARAMS);
+/* 1975C8 */ UNK_RET fn_801975C8(UNK_PARAMS);
+/* 1976D4 */ UNK_RET fn_801976D4(UNK_PARAMS);
+/* 1977AC */ UNK_RET fn_801977AC(UNK_PARAMS);
+/* 197AF0 */ UNK_RET fn_80197AF0(UNK_PARAMS);
+/* 197D4C */ UNK_RET fn_80197D4C(UNK_PARAMS);
+/* 197E18 */ UNK_RET fn_80197E18(UNK_PARAMS);
+/* 197FD8 */ UNK_RET fn_80197FD8(UNK_PARAMS);
+/* 1981A0 */ UNK_RET fn_801981A0(UNK_PARAMS);
+/* 1983E4 */ UNK_RET fn_801983E4(UNK_PARAMS);
+/* 198584 */ UNK_RET fn_80198584(UNK_PARAMS);
+/* 1985D4 */ UNK_RET fn_801985D4(UNK_PARAMS);
+/* 198824 */ UNK_RET fn_80198824(UNK_PARAMS);
+/* 198BA0 */ UNK_RET fn_80198BA0(UNK_PARAMS);
+/* 198C60 */ UNK_RET fn_80198C60(UNK_PARAMS);
+/* 198D18 */ UNK_RET fn_80198D18(UNK_PARAMS);
+/* 198EBC */ UNK_RET fn_80198EBC(UNK_PARAMS);
+/* 199AF0 */ UNK_RET fn_80199AF0(UNK_PARAMS);
+/* 19A158 */ UNK_RET fn_8019A158(UNK_PARAMS);
+/* 19A71C */ UNK_RET fn_8019A71C(UNK_PARAMS);
+/* 19A828 */ UNK_RET gm_8019A828(UNK_PARAMS);
+/* 19A86C */ UNK_RET fn_8019A86C(UNK_PARAMS);
+/* 19AF50 */ UNK_RET fn_8019AF50(UNK_PARAMS);
+/* 19B2DC */ UNK_RET gm_8019B2DC_OnFrame(UNK_PARAMS);
+/* 19B458 */ UNK_RET fn_8019B458(UNK_PARAMS);
+/* 19B81C */ UNK_RET fn_8019B81C(UNK_PARAMS);
+/* 19B860 */ UNK_RET fn_8019B860(UNK_PARAMS);
+
+#endif
