+++ src/melee/gm/gm_unsplit.h
@@ -1,0 +1,26 @@
+/**
+ * This is a helper file to reduce #include churn.
+ *
+ * If you need to include a function from an unsplit gm/ file,
+ * include this instead, since the unsplit file names are in flux.
+ *
+ * Once the function is moved to a properly split and named file,
+ * it won't be available via this header and you can use its proper include.
+ */
+
+#ifndef MELEE_GM_UNSPLIT_H
+#define MELEE_GM_UNSPLIT_H
+
+#include <melee/gm/gm_1601.h>
+#include <melee/gm/gm_17C0.h>
+#include <melee/gm/gm_1832.h>
+#include <melee/gm/gm_18A5.h>
+#include <melee/gm/gm_19EF.h>
+#include <melee/gm/gm_1A36.h>
+#include <melee/gm/gm_1A3F.h>
+#include <melee/gm/gm_1A45.h>
+#include <melee/gm/gm_1AED.h>
+#include <melee/gm/gm_1B03.h>
+#include <melee/gm/gm_1BFA.h>
+
+#endif
