+++ src/melee/gm/gmscdata.c
@@ -1,10 +1,6 @@
 #include "gmscdata.h"
 
-#include <melee/gm/gm_1601.h>
-#include <melee/gm/gm_19EF.h>
-#include <melee/gm/gm_1A4C.h>
-#include <melee/gm/gm_1AED.h>
-#include <melee/gm/gm_1B03.h>
+#include <melee/gm/gm_unsplit.h>
 #include <melee/gm/gmapproach.h>
 #include <melee/gm/gmcamera.h>
 #include <melee/gm/gmmenu.h>
